Feature/renderer (#11)
* Add Box2D as external dependency.

Also link it up to the engine.

* Fix imports

* Make 'game' compilable and runnable, with a dummy program

* Add CI support

* Add Box2D as external dependency.

Also link it up to the engine.

* Fix imports

* Make 'game' compilable and runnable, with a dummy program

* Add CI support

* Rebase to develop

* Remove remnant Box2D config from CMakeList

* Add 'remove' functionality to the managers

* Make game demo that uses current ECS implementation

* Add System and SystemManager

* Move game to 'demo/ecs' branch

* Add Box2D as external dependency.

Also link it up to the engine.

* Make 'game' compilable and runnable, with a dummy program

* Add CI support

* Make 'game' compilable and runnable, with a dummy program

* Add CI support

* Rebase to develop

* Add 'remove' functionality to the managers

* Make game demo that uses current ECS implementation

* Add System and SystemManager

* Move game to 'demo/ecs' branch

* Fix clang-format style issues

* Remove unused SDL imports in game

* Make IRenderer with SDL implementation. Make RenderVisitor with SDL implementation

* Fix CMake function

* Add LineShape

* Added Font graphics. Introduce new library (SDL_TTF)

* Update docs to reflect new library (SDL_TTF)

* Implement Sprite. Update docs to reflect new library

* Fix clang-format style issues

* Add new libs to CI. Fix remaining errors

* Fix more errors

+++ engine/window/SDLWindow.cpp
@@ -1,4 +1,5 @@
 #include "SDLWindow.h"
+
 #include <SDL.h>
 
 namespace engine {
@@ -41,7 +42,7 @@ int SDLWindow::getFlags() const
 int SDLWindow::getPosition(const int coord) const
 {
     if (m_properties.centered) {
-        return SDL_WINDOWPOS_CENTERED;
+        return static_cast<int>(SDL_WINDOWPOS_CENTERED);
     }
     return coord;
 }
